# SecurityService/src/main/java/com/romero/seguridad/auth/AuthenticationRequest.java
package com.romero.seguridad.auth;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class AuthenticationRequest {

  private String email;
  String password;
}