# SecurityService/src/main/java/com/romero/seguridad/auth/RegisterRequest.java
package com.romero.seguridad.auth;

import com.romero.seguridad.user.Role;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class RegisterRequest {

  private String firstname;
  private String lastname;
  private String email;
  private String password;
  private Role role;
}